iceberg-minio-demo/pyproject.toml
[project]
name = "iceberg-minio-demo"
version = "0.1.0"
description = "Learn Apache Iceberg with MinIO object storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Iceberg and data processing
    "pyiceberg[sql-sqlite,s3]>=0.9.0",
    "pandas>=2.3.2",
    "pyarrow>=17.0.0,<20.0.0",
    "duckdb>=1.3.2",
    
    # S3/MinIO connectivity
    "boto3>=1.35.0",
    "s3fs>=2024.6.0",
    
    # Configuration management
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    
    # CLI and utilities
    "click>=8.1.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PIE", "SIM", "RET"]
ignore = ["S101"]  # Allow assert statements in tests

[tool.black]
line-length = 100
target-version = ["py312"]